Integrate mesh subdivision based cutting algorithm
The cutting algorithm still does not handle transfering mass and velocity information to the subdivided tets. This will need to be done in order for the XPBD solver to correctly simulate the subdivided tets. The common case 4 seems to have a problem with topology where one of the vertices is not duplicated. Maybe this is just the shading that is misleading, but common case 4 should be investigated further to check for problems.