Implement transfer of physics quantities when cutting
Cutting modifies topology of the cut meshes, so new vertices are added, new adjacency information is created, and as such, masses, velocities and forces must be transfered to the new vertices, and physics constraints must be rebuilt.
We use linear interpolation to transfer velocities and forces to new vertices. However we still haven't decided how to transfer masses. Currently, we use linear interpolation, but mass distribution is uniform at vertices anyways, so the result is simply duplication of mass to the new vertices. We might want to design a proper mass redistribution scheme for this.