Refactor framework design
We establish clear separation between the rendering model and the physics model. The renderer only uses the render model. The solver uses the physics model, and might use some render models for collision detection. The render state now has three toggles, to minimize memory transfers to the GPU and to configure the type of rendering for individual scene nodes. For now, we support wireframe rendering. We also now make a clear separation between the physics mesh and the rendering surface mesh. Our scene now also provides clear separation between environment objects, which do not participate in deformations, and thus do not carry masses, weights, etc, but do participate in collision detection, and physics objects, which do participate in deformation.