Implement interactive particle fixing
We modify the mass transfer technique in the tetrahedron mesh cutter to account for fixed vertices. We always transfer the minimum mass surrounding vertex masses when creating new vertices upon cuts. In doing so, if the new vertex is surrounded by fixed particles, it will be fixed also. Otherwise, if there is one moving particle along its edge or along its face, then it will take the moving particle's mass. We also take the chance to fix a picking error. We extract the picked physics mesh's boundary surface mesh to perform ray intersection tests on. Otherwise, if we are in wireframe, we will test against each triangle of the physics mesh, including inner triangles.

With these changes, it is now possible to fix particles of a physics mesh, cut the mesh, and then run the simulation to observe the resulting behavior.

However, a major problem still occurs. In essence, when cutting a physics mesh composed of multiple mesh elements, those elements have adjacent faces and adjacent edges, which makes it such that when cutting one tetrahedron, the edge and face intersections we detect between the cutting tool and the mesh are shared between neighboring tetrahedra. However, since we cut each tetrahedron individually, we create duplicate vertices at the edge/face intersections, which incorrectly separates material at those points, where we will observe tets that should be glued together separate.