Upgrade tetrahedral mesh cutter for topology aware cutting
Progressive cutting still has bugs in it. The integration between cutting and the physics solver is not coherent. The progressive cutting creates new vertices at the moment of intersection detection, but topology only changes once the cutting surface exits the intersected tetrahedra. As such, the physics solver doesn't know about these new vertices, but updates all other vertices, so the new vertices hold physical quantities that were only valid in past frames.

To make the cutting topology aware, we now assign newly created vertices to the subdivided tetrahedra by checking the orientation of the cut edges. For each cut edge, there are 2 new vertices created with the same position, since material must separated at that edge. But how do we know which of the two vertices we assign to which subdivided elements? To cut each tetrahedron in a consistent manner, we now need to check which vertex end of a cut edge is "over" the cutting surface and which is "under" the cutting surface. In doing so, we can treat each common case generally by assigning "under" vertices to elements that are "under" the cut, and "over" vertices to elements that are "over" the cut.

There seem to be some cutting configurations which still are not handled correctly by the topology aware cutting algorithm. This must be investigated.

In summary, we still need to work on progressive cutting, since we need to pause the physics solver to perform cuts, such that they behave like "instantaneous" cuts. Also, we need to investigate if all cutting configurations are correctly supported.