Start the solver refactoring by exposing a new node interface
The scene will now only have a list of renderable nodes, which are interfaces to things that can generate triangle meshes. The load_scene function now takes factory methods that will actually construct the concrete classes implementing the renderable node interface. As such, we can now use any type of mesh as long as it implements the renderable node interface. In doing so, the solver can now use topological meshes instead of the shared vertex mesh data structure. The topological mesh data structure and the refactored xpbd solver system will help to implement constraint modifications when topological changes occur. The next developments will be to continue this refactoring. The topological mesh data structure needs to be implemented, and its refinable_mesh_adapter must wrap it and give it the ability to be cut by mesh subdivision, and have its corresponding constraints updated.