Implement dynamic mesh for environment body interactions in UI
A dynamic surface mesh type now is implemented and we can mutate its vertices and indices, such that they will be rendered to the scene updated. The trackball adapter has now been re-enabled for use with the new mesh system. It depends on a dynamic mesh, since it must modify the mesh to rotate and translate it. The pick function now also support twoway intersection tests so that triangle orientation does not matter in picking operations.