Implement St-Venant Kirchoff XPBD meshless method
Still much work to do. Imposition of Dirichlet boundary conditions does not work. Scaling the geometry of the meshless model does not seem to behave as in the mesh based techniques. I am guessing that this is because the coefficients at the meshless nodes do not satisfy the kronecker delta property. The stvk meshless beam example shows that the beam is initially shrunk, but then regains some volume, but the reason is unknown.