Implement particle grid sampling technique for meshless sph
Sample grid nodes at a given resolution inside the SDF of the given initial geometry for a meshless body. These grid nodes lying inside the SDF then become the SPH meshless nodes of the meshless body, as well as the SPH meshless particles in the XPBD simulation. Visual mesh rendering is done using the interpolation technique over the deformation gradient field and position field.

IMPORTANT:
The contact handling is not done correctly. When we detect a collision between some surface and the meshless body's visual surface, we get the meshless body's visual surface's vertex index. Then, we compute the mapping between this surface vertex index and its parent tetrahedral mesh's vertex. This is not correct, however, because the particles in the simulation are meshless nodes which do not correspond to the initial tet mesh's vertices. As such, we are creating collision constraints for the wrong particle indices (the tet mesh's vertex index). In reality, a collision with the surface mesh's vertex should be distributed to neighboring meshless particles, and a collision constraint must be generated for those neighboring particles. This is still unsolved/unimplemented in SBS.