Implement hybrid mesh/meshless model
This includes the mechanical model, composed of interior mesh nodal shape functions and meshless node shape functions around this interior mesh.
A visual model is also included, as well as the extended meshless nodes, which now are mixed particles, and must account for interior mesh nodes' shape functions in their interpolation scheme.

There remains to implement the mixed mesh/meshless particle constraint as well as how to correctly handle collisions. The handling of collisions is still a big problem. Visual model updating also must account for the interpolation along interior mesh node shape functions, not only the meshless node shape functions. This problem is avoided by having the surface vertices be the same as the tetrahedral mesh boundary vertices. However, this will have to change in the future.