Implement correct physical model initialization for hybrid mesh/meshless models
The meshless node to parent tetrahedron association was not done right previously. Now, we use a BVH over tetrahedra to correctly test for containment of meshless nodes in tets.

Also, the correct boundary surface extraction was done for the interior layer of the initial tetrahedral mesh geometry when constructing the hybrid mesh/meshless model. Because this boundary surface extraction is correctly oriented with outward pointing normals, we are able to build the right SDF for it, thus enabling correct particle sampling.

Other bugs were fixed related to indexing and uninitialized data to make the simulation run.

Now, collision detection is not working correctly. The contact handling does not handling meshless nodes and as such, user interaction is not possible. Even particle fixing does not work. This has to be fixed in further commits.