Implement hybrid mesh/meshless coupling with collisions
The implementation is not working. The coupling in the mixed domain gives bad deformation gradients. I think the solution would be to use first-order MLS shape functions with the requirement of continuity at the boundary for this to work. The moment matrix will make sure the MLS weights accommodate the mesh shape functions correctly.