Create math module
A math module with autodiff type interfaces was created to ease the process of creating experiments with different models with minimal effort. Different basis functions can be implemented, different domains, different fem models, etc. and everything can be differentiated automatically, which should be a huge productivity booster for future experiments.